Use RoundTripper-like interface for interceptors.
This simplifies the interceptor interface to look like
http.RoundTripper, having interceptors take in an http.Request and
returning an http.Response.

The only substantial behavior change this makes is gives webhook interceptors
the ability to respond back with HTTP headers.

Fixes #269.
