+++ pkg/interceptors/cel/cel_test.go
@@ -2,6 +2,7 @@ package cel
 
 import (
 	"bytes"
+	"io"
 	"io/ioutil"
 	"net/http"
 	"reflect"
@@ -135,16 +136,26 @@ func TestInterceptor_ExecuteTrigger(t *testing.T) {
 			}
 			request := &http.Request{
 				Body: ioutil.NopCloser(bytes.NewReader(tt.args.payload)),
+				GetBody: func() (io.ReadCloser, error) {
+					return ioutil.NopCloser(bytes.NewReader(tt.args.payload)), nil
+				},
 				Header: http.Header{
 					"Content-Type": []string{"application/json"},
 					"X-Test":       []string{"test-value"},
 				},
 			}
-			got, err := w.ExecuteTrigger(tt.args.payload, request, nil, "")
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Interceptor.ExecuteTrigger() error = %v, wantErr %v", err, tt.wantErr)
+			resp, err := w.ExecuteTrigger(request)
+			if err != nil {
+				if !tt.wantErr {
+					t.Errorf("Interceptor.ExecuteTrigger() error = %v, wantErr %v", err, tt.wantErr)
+				}
 				return
 			}
+			got, err := ioutil.ReadAll(resp.Body)
+			if err != nil {
+				t.Fatalf("error reading response body: %v", err)
+			}
+			defer resp.Body.Close()
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("Interceptor.ExecuteTrigger() = %s, want %s", got, tt.want)
 			}
