+++ pkg/interceptors/gitlab/gitlab.go
@@ -46,7 +46,8 @@ func NewInterceptor(gl *triggersv1.GitLabInterceptor, k kubernetes.Interface, ns
 	}
 }
 
-func (w *Interceptor) ExecuteTrigger(payload []byte, request *http.Request, _ *triggersv1.EventListenerTrigger, _ string) ([]byte, error) {
+//func (w *Interceptor) ExecuteTrigger(payload []byte, request *http.Request, _ *triggersv1.EventListenerTrigger, _ string) ([]byte, error) {
+func (w *Interceptor) ExecuteTrigger(request *http.Request) (*http.Response, error) {
 	// Validate the secret first, if set.
 	if w.GitLab.SecretRef != nil {
 		header := request.Header.Get("X-GitLab-Token")
@@ -78,5 +79,8 @@ func (w *Interceptor) ExecuteTrigger(payload []byte, request *http.Request, _ *t
 		}
 	}
 
-	return payload, nil
+	return &http.Response{
+		Header: request.Header,
+		Body:   request.Body,
+	}, nil
 }
